🔧 refactor: (#4) rename class
