+++ src/main/kotlin/entry/dsm/equusrecruitment/domain/application/exception/ApplicationAlreadyExistException.kt
@@ -1,6 +1,6 @@
 package entry.dsm.equusrecruitment.domain.application.exception
 
-import entry.dsm.equusrecruitment.domain.global.error.exception.EntryBlogException
 import entry.dsm.equusrecruitment.domain.global.error.exception.ErrorCode
+import entry.dsm.equusrecruitment.domain.global.error.exception.RecruitmentException
 
-object ApplicationAlreadyExistException : EntryBlogException(ErrorCode.APPLICATION_ALREADY_EXIST)
+object ApplicationAlreadyExistException : RecruitmentException(ErrorCode.APPLICATION_ALREADY_EXIST)
