+++ src/main/kotlin/entry/dsm/equusrecruitment/domain/global/error/exception/RecruitmentException.kt
